server/requirements.txt
Flask==2.2.5
requests==2.31.0
Pillow==9.5.0
google-genai==0.4.0
